perf(chat): pg_trgm search, per-tool rate limits, chat-tools split, file-budget guard
Closes the post-#146 review punch list:

* migration 019_pg_trgm_search — installs pg_trgm in `extensions`,
  adds GIN trigram indexes on grows.name|description and
  plants.name|strain|batch_label, and exposes SECURITY INVOKER
  search_grows / search_plants RPCs that rank by similarity score.
  Drop-in ILIKE acceleration plus better ranking for the agent's
  find_grow / find_plant tools.
* chat-tools.ts: switches find_grow / find_plant to the new RPCs
  with a defensive ILIKE fallback (so the agent stays functional
  if the migration hasn't been applied), adds a 45s timeout guard
  around trigger_plant_analysis so a stalled Gemini call doesn't
  hold the chat stream open, extracts CHAT_TOOL_DEFINITIONS into
  its own data module, and pulls per-tool rate-limit policy into
  chat-tool-policies.ts.
* per-tool rate limits: defence-in-depth budgets for
  trigger_plant_analysis (6/min), create_plants (10/min),
  create_grow (10/min), record_image_finding (20/min) — the chat
  route's overall ceiling stays, this caps single tools.
* plant-form: adopts the existing useActionWithRecovery hook so
  transient network errors retry once and router.push failures
  can't strand the user.
* grows constants: deduplicates MAX_DESCRIPTION_LENGTH and
  MAX_FUTURE_START_MS that PR #145 left split across new + [id].
* check-file-budgets.mjs: new guard wired into `pnpm run validate`
  so chat-tools.ts and chat-tool-definitions.ts can't silently
  grow past their refactor thresholds.

Verification: pnpm run validate, type-check, test (676/676), lint,
pytest (86/86), security:routes — all green.

Action required after merge: apply migration 019 to prod via
`supabase db push`. Until then the RPC path returns "function
search_grows does not exist" and the code path falls back to ILIKE.
